:sparkles: BE/Feat [S08P31A702-212] : Token Refresh 구현

See merge request s08-final/S08P31A702!116

+++ backend-business/FEELINGFILLING/src/main/java/com/a702/feelingfilling/domain/chatting/model/entity/Sender.java
@@ -29,7 +29,7 @@ public class Sender {
   private int senderId;
   @DBRef
   private List<Chatting> chattings;
-  private long numOfChat;
+  private int numOfChat;
   private int numOfUnAnalysed;
   private LocalDate lastDate;
 }
